move data from main to components
