+++ src/app/page.tsx
@@ -1,22 +1,8 @@
-"use client";
-
-import { useQuery } from "@tanstack/react-query";
-import { BookMinus, ChevronRight, Loader } from "lucide-react";
-import Image from "next/image";
-import Link from "next/link";
-import { Stats } from "./Stats";
-import { BookView } from "./BookView";
+import { Books } from "@/components/main/books";
+import { Profile } from "@/components/main/profile";
+import { Stats } from "@/components/main/stats";
 
 export default function Home() {
-  const booksQuery = useQuery({
-    queryKey: ["books"],
-    queryFn: () => fetch("/api/books").then((res) => res.json()),
-  });
-
-  const userQuery = useQuery({
-    queryKey: ["user"],
-    queryFn: () => fetch("/api/profile").then((res) => res.json()),
-  });
 
   return (
     <div>
@@ -24,84 +10,9 @@ export default function Home() {
         <div className="font-semibold mx-auto">Главная</div>
       </div>
       <div className="flex flex-col">
-        <div className="flex flex-col gap-3 border-b border-zinc-300 p-3 cursor-default">
-          <Link href="/books">
-            <h2 className="text-5xl font-black uppercase flex gap-1 items-center cursor-pointer hover:text-black/70 active:scale-90 transition-transform w-fit flex-wrap">
-              Книги
-              <ChevronRight className="w-12 h-12" strokeWidth={3} />
-            </h2>
-          </Link>
-          {booksQuery.isPending ? (
-            <div className="flex h-[20vh] justify-center items-center">
-              <Loader className="w-6 h-6 animate-spin" />
-            </div>
-          ) : (
-            <div className="flex flex-col gap-2">
-              {booksQuery?.data.slice(0, 3).map((book: Book) => (
-                <BookView book={book} key={book.id} />
-              ))}
-            </div>
-          )}
-          {booksQuery?.data?.length === 0 && (
-            <div className="p-2 flex gap-2 items-center rounded-xl border border-zinc-200 text-xl">
-              <BookMinus className="w-10 h-10" />
-              <div className="flex flex-col">
-                <div>Нет книг</div>
-                <div className="text-sm text-blue-500 underline underline-offset-4">
-                  <Link href="/books">Добавить книгу</Link>
-                </div>
-              </div>
-            </div>
-          )}
-        </div>
-        <div className="flex flex-col gap-3 border-b border-zinc-300 p-3 cursor-default">
-          <Link href="/profile#stats">
-            <h2 className="text-5xl font-black uppercase flex gap-1 items-center cursor-pointer hover:text-black/70 active:scale-90 transition-transform w-fit flex-wrap">
-              Статистика
-              <ChevronRight className="w-12 h-12" strokeWidth={3} />
-            </h2>
-          </Link>
-          <Stats />
-        </div>
-        <div className="flex flex-col gap-3 border-b border-zinc-300 p-3 cursor-default">
-          <Link href="/profile">
-            <h2 className="text-5xl font-black uppercase flex gap-1 items-center cursor-pointer hover:text-black/70 active:scale-90 transition-transform w-fit flex-wrap">
-              Профиль
-              <ChevronRight className="w-12 h-12" strokeWidth={3} />
-            </h2>
-          </Link>
-          {userQuery.isPending ? (
-            <div className="p-2 rounded-md border border-zinc-200 flex gap-2 items-center justify-center py-5">
-              <Loader className="w-6 h-6 animate-spin" />
-            </div>
-          ) : (
-            <div className="p-2 rounded-md border border-zinc-200 flex gap-2 items-center">
-              <Image
-                src={
-                  userQuery.data?.avatarUrl
-                    ? userQuery.data?.avatarUrl
-                    : "/no-avatar.png"
-                }
-                alt="avatar"
-                width={100}
-                height={100}
-                className="rounded-full w-20 h-20"
-              />
-              <div className="flex flex-col">
-                <div className="text-3xl font-semibold">
-                  {userQuery.data?.firstName} {userQuery.data?.lastName}
-                </div>
-                <div className="text-sm text-black/70">
-                  @{userQuery.data?.username}
-                </div>
-                {/* <div className="bg-blue-500 p-2 rounded-md shadow-md shadow-blue-300 cursor-pointer flex items-center justify-center text-white text-sm mt-2">
-            <UserPlus className="w-4 h-4 mr-2" />
-            Добавить в друзья
-          </div> */}
-              </div>
-            </div>
-          )}
-        </div>
+        <Books />
+        <Stats />
+        <Profile />
       </div>
     </div>
   );
